lambda/upload_code/requirements.txt
langchain==0.2.2
requests==2.25.1
GitPython==3.1.18
boto3==1.34.119